requirement.txt
Django==2.2
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.2
Pillow==6.0.0
psycopg2==2.8.2
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
